+++ frameworks/BJLiveCore.framework/Versions/A/Headers/BJLBaseVM.h
@@ -7,11 +7,13 @@
 //
 
 #import <Foundation/Foundation.h>
+#import <BJLiveBase/BJL_EXTScope.h>
 #import <BJLiveBase/BJL_M9Dev.h>
 #import <BJLiveBase/NSObject+BJL_M9Dev.h>
+#import <BJLiveBase/NSObject+BJLObserving.h>
 
 #import "BJLConstants.h"
-#import "BJLBlockHelper.h"
+#import "NSError+BJLError.h"
 
 NS_ASSUME_NONNULL_BEGIN